chore: change helper function to service function
